+++ src/import.rs
@@ -7,35 +7,56 @@ use crate::{
     args::import::Import,
     error::Result,
     specs::{
-        body::{Mapping, body::Body},
+        body::{Mapping, body::Body, type_constraint::TypeConstraint},
         method::Method,
-        response::Response,
+        mock_config::MockConfig,
+        response::{EndpointResponse, Response},
         spec::Spec,
-        specs_struct::Specs,
         status_code::StatusCode,
     },
 };
 
 impl Import {
-    pub fn oas3_to_specs(oas: OpenApiV3Spec) -> Result<Specs> {
+    pub fn oas3_to_specs(oas: OpenApiV3Spec) -> Result<MockConfig> {
         let Some(paths) = oas.paths.clone() else {
-            return Ok(Specs::default());
+            return Ok(MockConfig::default());
         };
 
-        let mut specs = Specs::default();
+        let mut specs = MockConfig::default();
         for (url, item) in paths.iter() {
             for (method, op) in item.methods() {
+                let req = Self::oper_to_req(op, &oas)?;
                 let res = Self::oper_to_res(op, &oas)?;
+
                 let spec = Spec {
                     method: Method::try_from(method)?,
                     url: url.clone(),
-                    response: res,
+                    request: req,
+                    response: EndpointResponse::Single(res),
                 };
-                specs.0.push(spec);
+                specs.specs.push(spec);
             }
         }
 
         Ok(specs)
+    }
+
+    fn oper_to_req(
+        op: &Operation,
+        oas: &OpenApiV3Spec,
+    ) -> Result<Option<Body>> {
+        let Some(req) = &op.request_body else {
+            return Ok(None);
+        };
+
+        let body = req.resolve(oas)?;
+        if let Some(media) = body.content.get("application/json")
+            && let Some(schema) = media.schema(oas)?
+        {
+            let body = Self::traverse_object(&schema, oas, true)?;
+            return Ok(Some(body));
+        }
+        Ok(None)
     }
 
     fn oper_to_res(op: &Operation, oas: &OpenApiV3Spec) -> Result<Response> {
@@ -48,7 +69,7 @@ impl Import {
         if let Some(media) = res.content.get("application/json")
             && let Some(schema) = media.schema(oas)?
         {
-            body = Self::traverse_object(&schema, oas)?;
+            body = Self::traverse_object(&schema, oas, false)?;
         }
 
         Ok(Response {
@@ -61,6 +82,7 @@ impl Import {
     fn traverse_object(
         obj: &ObjectSchema,
         spec: &OpenApiV3Spec,
+        is_req: bool,
     ) -> Result<Body> {
         let ty = match &obj.schema_type {
             Some(SchemaTypeSet::Single(ty)) => ty.clone(),
@@ -71,39 +93,67 @@ impl Import {
         };
 
         match ty {
-            SchemaType::Array => Self::parse_array(obj, spec),
-            SchemaType::Object => Self::parse_object(obj, spec),
-            SchemaType::String => Ok(Body::String(String::new())),
-            SchemaType::Integer => {
-                Ok(Body::Number(serde_yaml::Number::from(0)))
-            }
-            SchemaType::Number => {
-                Ok(Body::Number(serde_yaml::Number::from(0.0)))
-            }
-            SchemaType::Boolean => Ok(Body::Bool(true)),
-            SchemaType::Null => Ok(Body::Null),
+            SchemaType::Array => Self::parse_array(obj, spec, is_req),
+            SchemaType::Object => Self::parse_object(obj, spec, is_req),
+            SchemaType::String => Ok(if is_req {
+                Body::Constraint(TypeConstraint::new("string", None))
+            } else {
+                Body::String(String::new())
+            }),
+            SchemaType::Integer => Ok(if is_req {
+                Body::Constraint(TypeConstraint::new("number", None))
+            } else {
+                Body::Number(serde_yaml::Number::from(0))
+            }),
+            SchemaType::Number => Ok(if is_req {
+                Body::Constraint(TypeConstraint::new("number", None))
+            } else {
+                Body::Number(serde_yaml::Number::from(0.0))
+            }),
+            SchemaType::Boolean => Ok(if is_req {
+                Body::Constraint(TypeConstraint::new("boolean", None))
+            } else {
+                Body::Bool(true)
+            }),
+            SchemaType::Null => Ok(if is_req {
+                Body::Constraint(TypeConstraint::new("any", None))
+            } else {
+                Body::Null
+            }),
         }
     }
 
-    fn traverse_schema(schema: &Schema, spec: &OpenApiV3Spec) -> Result<Body> {
+    fn traverse_schema(
+        schema: &Schema,
+        spec: &OpenApiV3Spec,
+        is_req: bool,
+    ) -> Result<Body> {
         match schema {
-            Schema::Boolean(b) => Ok(Body::Bool(b.0)),
+            Schema::Boolean(b) => Ok(if is_req {
+                Body::Constraint(TypeConstraint::new("boolean", None))
+            } else {
+                Body::Bool(b.0)
+            }),
             Schema::Object(obj_ref) => {
                 let obj = obj_ref.resolve(spec)?;
-                Self::traverse_object(&obj, spec)
+                Self::traverse_object(&obj, spec, is_req)
             }
         }
     }
 
-    fn parse_array(obj: &ObjectSchema, spec: &OpenApiV3Spec) -> Result<Body> {
+    fn parse_array(
+        obj: &ObjectSchema,
+        spec: &OpenApiV3Spec,
+        is_req: bool,
+    ) -> Result<Body> {
         if let Some(items) = &obj.items {
-            let body = Self::traverse_schema(items, spec)?;
+            let body = Self::traverse_schema(items, spec, is_req)?;
             Ok(Body::Sequence(vec![body]))
         } else if !obj.prefix_items.is_empty() {
             let mut items = vec![];
             for item in obj.prefix_items.iter() {
                 let item = item.resolve(spec)?;
-                items.push(Self::traverse_object(&item, spec)?);
+                items.push(Self::traverse_object(&item, spec, is_req)?);
             }
             Ok(Body::Sequence(items))
         } else {
@@ -111,11 +161,15 @@ impl Import {
         }
     }
 
-    fn parse_object(obj: &ObjectSchema, spec: &OpenApiV3Spec) -> Result<Body> {
+    fn parse_object(
+        obj: &ObjectSchema,
+        spec: &OpenApiV3Spec,
+        is_req: bool,
+    ) -> Result<Body> {
         let mut map = Mapping::new();
         for (key, prop) in obj.properties.iter() {
             let prop = prop.resolve(spec)?;
-            let body = Self::traverse_object(&prop, spec)?;
+            let body = Self::traverse_object(&prop, spec, is_req)?;
             map.insert(Body::String(key.clone()), body);
         }
         Ok(Body::Mapping(map))