+++ src/specs/response/response.rs
@@ -11,6 +11,7 @@ use crate::{
 };
 
 #[derive(Debug, Default, Clone, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
 pub struct Response {
     #[serde(default)]
     pub status: StatusCode,
@@ -24,8 +25,9 @@ impl Response {
     pub fn to_http_response(
         &self,
         vars: &HashMap<String, UrlVar>,
+        templates: &HashMap<String, Body>,
     ) -> Result<HyperRes> {
-        let body = self.expand_vars(&vars);
+        let body = self.expand_vars(&vars, templates);
         let body = serde_json::to_string(&body).unwrap_or("".into());
 
         hyper::Response::builder()
@@ -38,7 +40,8 @@ impl Response {
     pub fn expand_vars(
         &self,
         vars: &HashMap<String, UrlVar>,
+        templates: &HashMap<String, Body>,
     ) -> serde_yaml::Value {
-        self.body.resolve(vars)
+        self.body.resolve(vars, templates)
     }
 }